+++ code/Meerkat.Security/Security/Activities/Configuration/ConfigElementCollection.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Configuration;
 
-namespace Meerkat.Configuration
+namespace Meerkat.Security.Activities.Configuration
 {
     /// <summary>
     /// Provides a generic collection of <see cref="ConfigurationElement" />, implementing most necessary behaviour for client types.